Refactor symbolic execution
