+++ lib/src/detectors/mod.rs
@@ -1,12 +1,10 @@
 pub mod detector;
 pub mod functions_detector;
-pub mod inputs_generator_detector;
 pub mod statistics_detector;
 pub mod strings_detector;
 
 use crate::detectors::detector::Detector;
 use crate::detectors::functions_detector::FunctionsDetector;
-use crate::detectors::inputs_generator_detector::InputsGeneratorDetector;
 use crate::detectors::statistics_detector::StatisticsDetector;
 use crate::detectors::strings_detector::StringsDetector;
 
@@ -23,10 +21,5 @@ macro_rules! create_detectors {
 
 /// Returns a vector of all the instantiated detectors
 pub fn get_detectors() -> Vec<Box<dyn Detector>> {
-    create_detectors!(
-        FunctionsDetector,
-        StringsDetector,
-        StatisticsDetector,
-        InputsGeneratorDetector
-    )
+    create_detectors!(FunctionsDetector, StringsDetector, StatisticsDetector)
 }
